Redesign suggestForm logic, code improvements #53
