+++ src/pages/PreviewProductPage.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import { Redirect } from 'react-router-dom';
 import { makeStyles } from '@material-ui/core/styles';
 
-import ProductDetailsPage from './ProductDetailsPage';
+import ProductDetailsPage from './ProductDetailsPage/ProductDetailsPage';
 
 const useStyles = makeStyles({
   root: {
@@ -12,6 +12,16 @@ const useStyles = makeStyles({
     flexDirection: 'column',
   },
 });
+
+const propTypes = {
+  formFields: PropTypes.shape({
+    prodName: PropTypes.string,
+    prodUrl: PropTypes.string,
+    headImg: PropTypes.string,
+    shortDescription: PropTypes.string,
+    fullDescription: PropTypes.string,
+  }).isRequired,
+};
 
 const PreviewProductPage = ({ formFields }) => {
   const classes = useStyles();
@@ -28,14 +38,6 @@ const PreviewProductPage = ({ formFields }) => {
 };
 const mapStateToProps = (state) => state.form;
 
-PreviewProductPage.propTypes = {
-  formFields: PropTypes.shape({
-    prodName: PropTypes.string,
-    prodUrl: PropTypes.string,
-    headImg: PropTypes.string,
-    shortDescription: PropTypes.string,
-    fullDescription: PropTypes.string,
-  }).isRequired,
-};
+PreviewProductPage.propTypes = propTypes;
 
 export default connect(mapStateToProps)(PreviewProductPage);
