+++ src/pages/index.js
@@ -1,0 +1,4 @@
+export { default as HomePage } from './HomePage';
+export { default as PreviewProductPage } from './PreviewProductPage';
+export { default as ProductDetailsPage } from './ProductDetailsPage/ProductDetailsPage';
+export { default as SuggestProductPage } from './SuggestProductPage/SuggestProductPage';
